requirements.txt
xero-python>=10.0.0
facebook-business>=19.0.0
requests>=2.31.0
python-dotenv>=1.0.0
jinja2>=3.1.3
schedule>=1.2.1
